Depth Map Project
The purpose of this project was to simulate depth perception on a
computer. The program takes in two stereo images (one imaged from the
right, the other from the left). It was my job to implement the
functions in the file calc_depth.c and make_qtree.c. The calc_depth()
function does all the heavy lifting by converting the two input stereo
images into a depth map. The file make_qtree.c creates a quadtree -
convenient way to store the depth map images.